Support for MP and SP devices
Python-broadlink already has support for MP and SP (SP2 family) devices.
I have just added the type field to select the right type of device (RM, SP, MP) and added the call for controlling the MP socket.
The code is tested with the MP2 device. 
I will test it and adopt for the MP1 device as soon as it arrives.