MP1 strip support
I have just started changing the code for MP1 support that is committed yesterday by python-broadlink project.
Still not sure how to use configuration file to pass Socket number as param (there are 4 independent sockets).
This version is working now with MP1 and socket 1 as this socket nr. is hard coded in check and set lines.